+++ backend/__tests__/unit/modules/github/controllers/getPullRequests.test.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express';
-import { getIssues } from '@/modules/github/controller';
+import { getPullRequests } from '@/modules/github/controllers';
 import { HttpError, httpStatusCode } from '@/utils';
 
 // Mock the Express request and response objects
@@ -13,10 +13,10 @@ const nextFunction = jest.fn();
 // Mock the console.error function
 console.error = jest.fn();
 
-// Test suite for getIssues controller
-describe('getIssues controller', () => {
-  // Test case: should fetch issues from GitHub API successfully
-  it('fetch issues from GitHub API successfully', async () => {
+// Test suite for getPullRequests controller
+describe('getPullRequests controller', () => {
+  // Test case: should fetch pull requests from GitHub API successfully
+  it('fetch pull requests from GitHub API successfully', async () => {
     // Mock the getZodValidatedParams and getZodValidatedQuery methods
     req.getZodValidatedParams = jest.fn().mockReturnValue({ repo: 'developer-experience-team' });
     req.getZodValidatedQuery = jest.fn().mockReturnValue({ state: 'open' });
@@ -26,9 +26,10 @@ describe('getIssues controller', () => {
       ok: true,
       json: jest.fn().mockResolvedValue([
         {
-          html_url: 'https://github.com/bcgov/developer-experience-team/issues/1',
+          html_url: 'https://github.com/bcgov/developer-experience-team/pull/1',
           number: 1,
-          title: 'Example Issue',
+          pull_request: true,
+          title: 'Example Pull Request',
           comments: 2,
           created_at: '2024-04-09T12:00:00Z',
           updated_at: '2024-04-09T13:00:00Z',
@@ -44,14 +45,14 @@ describe('getIssues controller', () => {
     global.fetch = jest.fn().mockResolvedValue(mockResponse);
 
     // Call the controller function
-    await getIssues(req, res, nextFunction);
+    await getPullRequests(req, res, nextFunction);
 
     // Assert that the response was sent with the correct data
     expect(res.json).toHaveBeenCalledWith([
       {
-        url: 'https://github.com/bcgov/developer-experience-team/issues/1',
+        url: 'https://github.com/bcgov/developer-experience-team/pull/1',
         number: 1,
-        title: 'Example Issue',
+        title: 'Example Pull Request',
         comments: 2,
         created_at: '2024-04-09T12:00:00Z',
         updated_at: '2024-04-09T13:00:00Z',
@@ -79,7 +80,7 @@ describe('getIssues controller', () => {
     global.fetch = jest.fn().mockResolvedValue(mockErrorResponse);
 
     try {
-      await getIssues(req, res, nextFunction);
+      await getPullRequests(req, res, nextFunction);
     } catch (error) {
       expect(error).toBeInstanceOf(HttpError);
       expect((error as HttpError).statusCode).toBe(httpStatusCode.BAD_REQUEST);
@@ -102,7 +103,7 @@ describe('getIssues controller', () => {
     global.fetch = jest.fn().mockResolvedValue(mockResponse);
 
     try {
-      await getIssues(req, res, nextFunction);
+      await getPullRequests(req, res, nextFunction);
     } catch (error) {
       expect(error).toBeInstanceOf(HttpError);
       expect((error as HttpError).statusCode).toBe(httpStatusCode.NOT_FOUND);
