+++ backend/src/modules/github/controllers/index.ts
@@ -1,0 +1,2 @@
+export * from './getIssues';
+export * from './getPullRequests';
